fix(extension): do not chain txBuilder calls
TxBuilder is a proxy object in the extension. Chaining it's calls
creates new proxy objects at each step. This seems to mess up the
service worker <-> extension communication.
Using the initial TxBuilder proxy object, executes all calls on
the same proxy object, and this seems to work better to preserve the
keyAgent message passing.
Please note the number of "seems" word usages... indicating that the
root cause is not yet determined.